+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Gitleaks test report generated from Juice Shop v14.3.0-4-g2c757a928
 - Gitleaks to CLI
 - Gitleaks as Export target to Defect Dojo
+- Blacklist Validation using KEVs from CISA
 
 
 ### Changed